+++ Sources/PerformanceAnalyst/__init__.py
@@ -5,6 +5,8 @@ PerformanceAnalyst
 This package is for timing pieces of code and keeping track of the timings over time for postprocessing. The design and terminology is based on Python's `unit testing framework <http://docs.python.org/library/unittest.html#module-unittest>`_, so it may be useful to refer to its documentation too.
 """
 
+from _Configuration import *
+
 import CompositeTimerRunner
 import ProgressTimerRunner
 import SQLiteTimerRunner