+++ Documentation/History.rst
@@ -8,7 +8,8 @@ History
 ======================
 - Added SQLiteTimerRunner.databaseIsConsistent.
 - Added support for tracking real time, besides CPU-time. This required a change in the database structure. See the :py:mod:`SQLiteTimerRunner's warnings <PerformanceAnalyst.SQLiteTimerRunner>`!
-- The `pa.py` ls command now orders the per-timer results by timestamp.
+- The `pa.py ls` command now orders the per-timer results by timestamp.
+- The `pa.py plot` command not has a `timestamp` option which is used to select timestamps to use for plot. Data created before this timestamp is skipped.
 
 0.0.5
 =====